refactor(Command): 패키지 이동(domain -> view)
